+++ Accounts.Repository/Repository/IAccountTransMasterRepository.cs
@@ -12,7 +12,7 @@ namespace Accounts.Repository.Repository
     public interface IAccountTransMasterRepository
     {
         List<AccountTransMaster> GetAllAccountTransMaster();
-        AccountTransMaster FindAccountTransMaster(int id);
+        AccountTransMaster FindAccountTransMaster(long id);
         bool UpdateAccountTransMaster(VM_AccountTransMaster _VM_AccountTransMaster);
         bool DeleteAccountTransMaster(int id);
         bool AddAccountTransMaster(VM_AccountTransMaster _VM_AccountTransMaster);
